hackday2020_backend/requirements.txt
Flask==1.1.2
Pillow==7.1.2
flask-api==2.0
loguru==0.5.1